moved useFirebase
